Revert "Merge branch 'main' into main"
This reverts commit dbf091bb8115ad388d8af7a32e9f0a71d5aa6f5b, reversing
changes made to f3d81698f521ac512ad266144afcfa3e72a6347e.

+++ FluidNC/src/Stepping.h
@@ -38,7 +38,7 @@ namespace Machine {
         // execution lead time there is for other processes to run.  The latency for a feedhold or other
         // override is roughly 10 ms times _segments.
 
-        int32_t _segments = 12;
+        size_t _segments = 12;
 
         uint32_t _idleMsecs           = 255;
         uint32_t _pulseUsecs          = 4;
